+++ src/deact/elements/Fragement.js
@@ -1,0 +1,8 @@
+// TODO: children
+const Fragment = async () => ({
+  contents: [],
+  embeds: [],
+  components: [],
+});
+
+module.exports = Fragment;