+++ types.js
@@ -62,6 +62,10 @@
  *  money?: number,
  *  backpack?: FlattenedBackpack
  * }} FlattenedRewards
+ *
+ * These are the same type for now
+ * @typedef {Rewards} Cost
+ * @typedef {FlattenedRewards} FlattenedCost
  */
 
 /**
@@ -171,6 +175,7 @@
  * Important info (IDK)
  * @property {NatureEnum} natureId
  * @property {AbilityIdEnum | string | number} abilityId
+ * @property {HeldItemIdEnum?} heldItemId
  * @property {boolean} shiny
  * @property {MoveIdEnum[]} moveIds
  *